Change heap cost type for better integration with hashmap
git-svn-id: https://gnunet.org/svn/GNUnet@8767 140774ce-b5e7-0310-ab8b-a85725594a96

+++ src/include/gnunet_util_containers.h
@@ -578,7 +578,7 @@ void *GNUNET_multi_hash_map_get_random (const struct GNUNET_MultiHashMap
     (element)->next->prev = (element)->prev;
 
 
-typedef unsigned int GNUNET_CostType;
+typedef unsigned long long GNUNET_CostType;
 
 /*
  * Heap type, either max or min.  Hopefully makes the
